requirements.txt
Django==3.2.10
psycopg2>=2.8
djangorestframework==3.12.4
